refactor(filter dialogs): Change file names and component names for homogeneity Fix GenericFilterDialog imported name which was still used and handleCloseGenericFilterDialog replaced by handleCloseCriteriaBasedFilterDialog Rename ExplicitNamingCreationDialog to ExplicitNamingFilterCreationDialog Rename file criteria-filter-* to criteria-based-filter-* to match component name
