+++ src/components/dialogs/criteria-based-filter-dialog-content.js
@@ -81,7 +81,7 @@ export const FilterTypeSelection = ({
     );
 };
 
-export const CriteriaFilterDialogContent = ({
+export const CriteriaBasedFilterDialogContent = ({
     id,
     open,
     contentType,
@@ -255,4 +255,4 @@ export const CriteriaFilterDialogContent = ({
     );
 };
 
-export default CriteriaFilterDialogContent;
+export default CriteriaBasedFilterDialogContent;
