+++ src/components/dialogs/explicit-naming-filter-dialog-content.js
@@ -20,7 +20,7 @@ import { Draggable } from 'react-beautiful-dnd';
 import PropTypes from 'prop-types';
 import { Alert, Checkbox, Input, Tooltip } from '@mui/material';
 import { filterEquipmentDefinition } from '../../utils/equipment-types';
-import { FilterTypeSelection } from './criteria-filter-dialog-content';
+import { FilterTypeSelection } from './criteria-based-filter-dialog-content';
 
 const useStyles = makeStyles(() => ({
     dialogPaper: {
